+++ lib/api/categories.js
@@ -1,0 +1,225 @@
+import fetchAPI from "@/lib/api/utils"
+
+export async function getPageBySlugAndCategory(slug) {
+  const data = await fetchAPI(
+    `query PageBySlug($where: JSON){
+      pages(where: $where) {
+        title_en
+        title {
+          title_es
+          title_tr
+        }
+        slug
+        description {
+          description_en
+          description_es
+          description_tr
+        }
+        body {
+          __typename
+          ... on ComponentContentContent {
+            content_en
+            content_es
+            content_es
+          }
+          __typename
+          ... on ComponentContentExperience {
+            id
+            from
+            to
+            ongoing
+            title
+            institution
+            address {
+              address
+              city
+              country
+              postalcode
+            }
+            url
+            description{
+              content_en
+              content_es
+              content_tr
+            }
+          }
+          __typename
+          ... on ComponentContentPersonalInformation {
+            name
+            photo {
+              alternativeText
+              caption
+              url
+              previewUrl
+              width
+              height
+            }
+            position
+            nationality
+            address {
+              address
+              city
+              country
+              postalcode
+            }
+            telephone {
+              id
+              type
+              number
+            }
+            mail
+            links {
+              id
+              type
+              url
+            }
+            aboutme {
+              content_en
+              content_es
+              content_tr
+            }
+          }
+          __typename
+          ... on ComponentFieldsSkill {
+            type
+            id
+            name
+            level
+          }
+          __typename
+          ... on ComponentHeaderCategory {
+            name
+            description{
+              description_en
+              description_es
+              description_tr
+            }
+            projects {
+              id
+              title
+              description {
+                content_en
+                content_es
+                content_tr
+              }
+              thumbnail {
+                alternativeText
+                url
+              }
+            }
+          }
+        }
+        tags{
+          id
+          label
+          slug
+        }
+      }
+      categories {
+        id
+        slug
+        name
+        description {
+          description_en
+          description_es
+          description_tr
+        }
+        projects {
+          id
+          title
+          thumbnail {
+            alternativeText
+            caption
+            previewUrl
+            width
+            height
+          }
+          description {
+            content_en
+            content_es
+            content_tr
+          }
+        }
+      }
+    }`,
+    {
+      variables: {
+        where: { slug },
+      },
+    }
+  )
+  return data
+}
+
+export async function getAllCategories() {
+  const data = await fetchAPI(`{
+      categories {
+        id
+        slug
+        name
+        description {
+          description_en
+          description_es
+          description_tr
+        }
+        projects {
+          id
+          title
+          thumbnail {
+            alternativeText
+            caption
+            previewUrl
+            width
+            height
+          }
+          description {
+            content_en
+            content_es
+            content_tr
+          }
+        }
+      }
+    }
+    `)
+  return data?.allCategories
+}
+
+export async function getCategoryBySlug(slug) {
+  const data = await fetchAPI(
+    `query CategoryBySlug($where: JSON){
+      categories(where: $where){
+        id
+        slug
+        name
+        description{
+          description_en
+          description_es
+          description_tr
+        }
+        projects{
+          id
+          title
+          thumbnail{
+            alternativeText
+            caption
+            previewUrl
+            width
+            height
+          }
+          description{
+            content_en
+            content_es
+            content_tr
+          }
+        }
+      }
+    }`,
+    {
+      variables: {
+        where: { slug },
+      },
+    }
+  )
+
+  return data
+}
