+++ lib/api/posts.js
@@ -1,0 +1,3 @@
+import fetchAPI from "@/lib/api/utils"
+
+export async function getBlogs() {}
