+++ pages/_app.js
@@ -6,7 +6,7 @@ import { ThemeProvider, CssBaseline } from "@material-ui/core"
 import { appWithTranslation } from "i18n"
 import Meta from "../components/meta"
 import DefaultSEO from "../next-seo.config"
-import { getPageTitleBySlug, getPageTitlesBySlugSet } from "@/lib/api"
+import { getPageTitlesBySlugSet } from "@/lib/api/pages"
 import { CMS_TILE_COLOR, menu_links } from "@/lib/constants"
 import theme from "@/styles/theme"
 import Nav from "@/components/nav"
