+++ pages/about-me.js
@@ -5,7 +5,7 @@ import { SocialProfileJsonLd, NextSeo } from "next-seo"
 import { useTranslation } from "react-i18next"
 import { makeStyles, Typography } from "@material-ui/core"
 import { Skeleton } from "@material-ui/lab"
-import { getPageBySlug } from "@/lib/api"
+import { getPageBySlug } from "@/lib/api/pages"
 import { CMS_URL, CMS_NAME } from "@/lib/constants"
 import SEO from "@/components/seo"
 import SkillsSection from "@/components/skillssection"
