+++ pages/education.js
@@ -4,7 +4,7 @@ import { useRouter } from "next/router"
 import { useTranslation } from "react-i18next"
 import { Typography } from "@material-ui/core"
 import { Skeleton, Timeline } from "@material-ui/lab"
-import { getPageBySlug } from "@/lib/api"
+import { getPageBySlug } from "@/lib/api/pages"
 import Layout from "@/components/layout"
 import SEO from "@/components/seo"
 import Experience from "@/components/content/experience"
