+++ pages/work-experience.js
@@ -2,7 +2,7 @@ import PropTypes from "prop-types"
 import ErrorPage from "next/error"
 import { useRouter } from "next/router"
 import { useTranslation } from "react-i18next"
-import { getPageBySlug } from "@/lib/api"
+import { getPageBySlug } from "@/lib/api/pages"
 import Layout from "@/components/layout"
 import { Skeleton } from "@material-ui/lab"
 import SEO from "@/components/seo"
